+++ kafka-connect-aws-s3/src/main/scala/io/lenses/streamreactor/connect/aws/s3/formats/S3FormatWriter.scala
@@ -17,11 +17,10 @@
 
 package io.lenses.streamreactor.connect.aws.s3.formats
 
-import io.lenses.streamreactor.connect.aws.s3.Topic
 import io.lenses.streamreactor.connect.aws.s3.config.Format._
 import io.lenses.streamreactor.connect.aws.s3.config.FormatOptions.WithHeaders
 import io.lenses.streamreactor.connect.aws.s3.config.{BytesWriteMode, FormatOptions, FormatSelection}
-import io.lenses.streamreactor.connect.aws.s3.model.SinkData
+import io.lenses.streamreactor.connect.aws.s3.model.{SinkData, Topic}
 import io.lenses.streamreactor.connect.aws.s3.storage.MultipartBlobStoreOutputStream
 
 object S3FormatWriter {
