+++ kafka-connect-aws-s3/src/test/scala/io/lenses/streamreactor/connect/aws/s3/sink/CommittedFileNameTest.scala
@@ -19,8 +19,7 @@ package io.lenses.streamreactor.connect.aws.s3.sink
 
 import io.lenses.streamreactor.connect.aws.s3.config.Format.{Avro, Csv, Json}
 import io.lenses.streamreactor.connect.aws.s3.config.FormatSelection
-import io.lenses.streamreactor.connect.aws.s3.model.{PartitionSelection, ValuePartitionField}
-import io.lenses.streamreactor.connect.aws.s3.{Offset, Topic}
+import io.lenses.streamreactor.connect.aws.s3.model.{Offset, PartitionSelection, Topic, ValuePartitionField}
 import org.scalatest.flatspec.AnyFlatSpecLike
 import org.scalatest.matchers.should.Matchers
