+++ kafka-connect-aws-s3/src/test/scala/io/lenses/streamreactor/connect/aws/s3/sink/S3JsonWriterManagerTest.scala
@@ -17,10 +17,9 @@
 
 package io.lenses.streamreactor.connect.aws.s3.sink
 
-import io.lenses.streamreactor.connect.aws.s3._
 import io.lenses.streamreactor.connect.aws.s3.config.Format.Json
 import io.lenses.streamreactor.connect.aws.s3.config.{AuthMode, BucketOptions, FormatSelection, S3Config}
-import io.lenses.streamreactor.connect.aws.s3.model.{MessageDetail, StructSinkData}
+import io.lenses.streamreactor.connect.aws.s3.model.{BucketAndPrefix, MessageDetail, Offset, StructSinkData, Topic, TopicPartitionOffset}
 import io.lenses.streamreactor.connect.aws.s3.sink.utils.{S3ProxyContext, S3TestConfig}
 import org.apache.kafka.connect.data.Struct
 import org.jclouds.blobstore.options.ListContainerOptions
